+++ trianglemesh.cpp
@@ -120,20 +120,20 @@ void TriangleMesh::writeMeshToOBJ(std::string filename) {
     str << "# face count = " << triangles.size() << std::endl;
 
     glm::vec3 p;
-    for (int i = 0; i < vertices.size(); i++) {
+    for (int i = 0; i < (int)vertices.size(); i++) {
         p = vertices[i];
         str << "v " << p.x << " " << p.y << " " << p.z << std::endl;
     }
 
     glm::vec3 n;
-    for (int i = 0; i < normals.size(); i++) {
+    for (int i = 0; i < (int)normals.size(); i++) {
         n = normals[i];
         str << "vn " << n.x << " " << n.y << " " << n.z << std::endl;
     }
 
     Triangle t;
     int v1, v2, v3;
-    for (int i = 0; i < triangles.size(); i++) {
+    for (int i = 0; i < (int)triangles.size(); i++) {
         t = triangles[i];
         v1 = t.tri[0] + 1;
         v2 = t.tri[1] + 1;
@@ -167,7 +167,7 @@ void TriangleMesh::removeDuplicateTriangles() {
 
     std::sort(triangles.begin(), triangles.end(), triangleSort);
     Triangle last;
-    for (int i = 0; i < triangles.size(); i++) {
+    for (int i = 0; i < (int)triangles.size(); i++) {
         Triangle t = triangles[i];
 
         if (!_trianglesEqual(t, last)) {
@@ -185,10 +185,10 @@ void TriangleMesh::updateVertexNormals() {
     _updateVertexTriangles();
     
     std::vector<glm::vec3> facenormals;
-    facenormals.reserve(triangles.size());
+    facenormals.reserve((int)triangles.size());
     Triangle t;
     glm::vec3 v1, v2;
-    for (int i = 0; i < triangles.size(); i++) {
+    for (int i = 0; i < (int)triangles.size(); i++) {
         t = triangles[i];
 
         v1 = vertices[t.tri[1]] - vertices[t.tri[0]];
@@ -204,9 +204,9 @@ void TriangleMesh::updateVertexNormals() {
     }
 
     glm::vec3 n;
-    for (int i = 0; i < _vertexTriangles.size(); i++) {
+    for (int i = 0; i < (int)_vertexTriangles.size(); i++) {
         n = glm::vec3(0.0, 0.0, 0.0);
-        for (int j = 0; j < _vertexTriangles[i].size(); j++) {
+        for (int j = 0; j < (int)_vertexTriangles[i].size(); j++) {
             n += facenormals[_vertexTriangles[i][j]];
         }
 
@@ -232,7 +232,7 @@ bool TriangleMesh::_trianglesEqual(Triangle &t1, Triangle &t2) {
 bool TriangleMesh::isNeighbours(Triangle t1, Triangle t2) {
     std::vector<int> n;
     getFaceNeighbours(t1, n);
-    for (int i = 0; i < n.size(); i++) {
+    for (int i = 0; i < (int)n.size(); i++) {
         if (_trianglesEqual(triangles[n[i]], t2)) {
             return true;
         }
@@ -245,7 +245,7 @@ void TriangleMesh::_updateVertexTriangles() {
     _vertexTriangles.clear();
     _vertexTriangles.reserve(vertices.size());
 
-    for (int i = 0; i < vertices.size(); i++) {
+    for (int i = 0; i < (int)vertices.size(); i++) {
         std::vector<int> triangles;
         triangles.reserve(14);  // 14 is the maximum number of adjacent triangles
                                 // to a vertex
@@ -253,7 +253,7 @@ void TriangleMesh::_updateVertexTriangles() {
     }
 
     Triangle t;
-    for (int i = 0; i < triangles.size(); i++) {
+    for (int i = 0; i < (int)triangles.size(); i++) {
         t = triangles[i];
         _vertexTriangles[t.tri[0]].push_back(i);
         _vertexTriangles[t.tri[1]].push_back(i);
@@ -274,7 +274,7 @@ void TriangleMesh::_getTriangleGridCellOverlap(Triangle t, std::vector<GridIndex
     tbbox.getOverlappingGridCells(_dx, testcells);
 
     AABB cbbox = AABB(glm::vec3(0.0, 0.0, 0.0), _dx, _dx, _dx);
-    for (int i = 0; i < testcells.size(); i++) {
+    for (int i = 0; i < (int)testcells.size(); i++) {
         cbbox.position = _gridIndexToPosition(testcells[i]);
         if (cbbox.isOverlappingTriangle(t, vertices)) {
             cells.push_back(testcells[i]);
@@ -290,12 +290,12 @@ void TriangleMesh::_updateTriangleGrid() {
     std::vector<int> *triVector;
     Triangle t;
     GridIndex g;
-    for (int i = 0; i < triangles.size(); i++) {
+    for (int i = 0; i < (int)triangles.size(); i++) {
         t = triangles[i];
         cells.clear();
         _getTriangleGridCellOverlap(t, cells);
 
-        for (int j = 0; j < cells.size(); j++) {
+        for (int j = 0; j < (int)cells.size(); j++) {
             g = cells[j];
             triVector = _triGrid.getPointer(g);
             triVector->push_back(i);
@@ -373,7 +373,7 @@ void TriangleMesh::getTrianglePosition(unsigned int index, glm::vec3 tri[3]) {
     assert(index < triangles.size());
 
     Triangle t = triangles[index];
-    int size = vertices.size();
+    int size = (int)vertices.size();
     assert(t.tri[0] < size && t.tri[1] < size && t.tri[2] < size);
 
     tri[0] = vertices[t.tri[0]];
@@ -382,30 +382,30 @@ void TriangleMesh::getTrianglePosition(unsigned int index, glm::vec3 tri[3]) {
 }
 
 glm::vec3 TriangleMesh::getTriangleNormal(unsigned int index) {
-    assert(index < triangles.size());
+    assert(index < (int)triangles.size());
 
     Triangle t = triangles[index];
-    int size = vertices.size();
+    int size = (int)vertices.size();
     assert(t.tri[0] < size && t.tri[1] < size && t.tri[2] < size);
 
     return glm::normalize(normals[t.tri[0]] + normals[t.tri[1]] + normals[t.tri[2]]);
 }
 
 glm::vec3 TriangleMesh::getTriangleFaceDirection(unsigned int index) {
-    assert(index < triangles.size());
+    assert(index < (int)triangles.size());
 
     Triangle t = triangles[index];
-    int size = vertices.size();
+    int size = (int)vertices.size();
     assert(t.tri[0] < size && t.tri[1] < size && t.tri[2] < size);
 
     return normals[t.tri[0]] + normals[t.tri[1]] + normals[t.tri[2]];
 }
 
 glm::vec3 TriangleMesh::getTriangleCenter(unsigned int index) {
-    assert(index < triangles.size());
+    assert(index < (int)triangles.size());
 
     Triangle t = triangles[index];
-    int size = vertices.size();
+    int size = (int)vertices.size();
     assert(t.tri[0] < size && t.tri[1] < size && t.tri[2] < size);
 
     return (vertices[t.tri[0]] + vertices[t.tri[1]] + vertices[t.tri[2]]) / 3.0f;
@@ -426,7 +426,7 @@ bool TriangleMesh::_isOnTriangleEdge(double u, double v) {
 }
 
 bool TriangleMesh::_isTriangleInVector(int index, std::vector<int> &tris) {
-    for (int i = 0; i < tris.size(); i++) {
+    for (int i = 0; i < (int)tris.size(); i++) {
         if (_trianglesEqual(triangles[index], triangles[tris[i]])) {
             return true;
         }
@@ -453,7 +453,7 @@ int TriangleMesh::_getIntersectingTrianglesInCell(GridIndex g, glm::vec3 p, glm:
     int numIntersections = 0;
 
     numIntersections = 0;
-    for (int i = 0; i < indices->size(); i++) {
+    for (int i = 0; i < (int)indices->size(); i++) {
         getTrianglePosition(indices->at(i), tri);
 
         bool isIntersecting = Collision::lineIntersectsTriangle(p, dir, 
@@ -477,7 +477,7 @@ int TriangleMesh::_getIntersectingTrianglesInCell(GridIndex g, glm::vec3 p, glm:
 }
 
 bool TriangleMesh::_isIntInVector(int v, std::vector<int> &ints) {
-    for (int i = 0; i < ints.size(); i++) {
+    for (int i = 0; i < (int)ints.size(); i++) {
         if (ints[i] == v) {
             return true;
         }
@@ -522,7 +522,7 @@ bool TriangleMesh::_isCellInsideMesh(const GridIndex g) {
             return false;
         }
 
-        for (int i = 0; i < intersections.size(); i++) {
+        for (int i = 0; i < (int)intersections.size(); i++) {
             int idx = intersections[i];
             if (!_isIntInVector(idx, allIntersections) && !_isTriangleInVector(idx, allIntersections)) {
                 leftIntersections.push_back(idx);
@@ -544,7 +544,7 @@ bool TriangleMesh::_isCellInsideMesh(const GridIndex g) {
             return false;
         }
 
-        for (int i = 0; i < intersections.size(); i++) {
+        for (int i = 0; i < (int)intersections.size(); i++) {
             int idx = intersections[i];
             if (!_isIntInVector(idx, allIntersections) && !_isTriangleInVector(idx, allIntersections)) {
                 rightIntersections.push_back(idx);
@@ -579,7 +579,7 @@ void TriangleMesh::getCellsInsideMesh(std::vector<GridIndex> &cells) {
 
     GridIndex neighbours[6];
     GridIndex n;
-    for (int i = 0; i < surfaceCells.size(); i++) {
+    for (int i = 0; i < (int)surfaceCells.size(); i++) {
         _getNeighbourGridIndices6(surfaceCells[i], neighbours);
         for (int j = 0; j < 6; j++) {
             n = neighbours[j];
@@ -611,10 +611,10 @@ void TriangleMesh::_smoothTriangleMesh(double value) {
     glm::vec3 nv;
     glm::vec3 avg;
     Triangle t;
-    for (int i = 0; i < vertices.size(); i++) {
+    for (int i = 0; i < (int)vertices.size(); i++) {
 
         avg = glm::vec3(0.0, 0.0, 0.0);
-        for (int j = 0; j < _vertexTriangles[i].size(); j++) {
+        for (int j = 0; j < (int)_vertexTriangles[i].size(); j++) {
             t = triangles[_vertexTriangles[i][j]];
             if (t.tri[0] != i) {
                 avg += vertices[t.tri[0]];
